removing a few warning andunrolling a tensor loop

+++ src/adjmat/AdjacencyMatrixBase.h
@@ -374,7 +374,7 @@ void AdjacencyMatrixBase<T>::calculate() {
 
   // Reshape the matrix store if the number of columns has changed
   if( maxcol!=myval->getNumberOfColumns() ) {
-    for(int i=0; i<getNumberOfComponents(); ++i) {
+    for(size_t i=0; i<getNumberOfComponents(); ++i) {
       getPntrToComponent(i)->reshapeMatrixStore( maxcol );
     }
   }
